ER-785 - Moved Cache from Employer.Web to Client
