+++ src/Shared/Recruit.Vacancies.Client/Application/Cache/Cache.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using Microsoft.Extensions.Caching.Memory;
 
-namespace Esfa.Recruit.Employer.Web.Caching
+namespace Esfa.Recruit.Vacancies.Client.Application.Cache
 {
     public class Cache : ICache
     {
